Update main branch

+++ src/protagonist/API/Features/Space/Requests/DeleteSpace.cs
@@ -8,33 +8,24 @@ namespace API.Features.Space.Requests;
 /// <summary>
 /// Deletes a specified space for customer
 /// </summary>
-public class DeleteSpace: IRequest<ResultMessage<DeleteResult>>
+public class DeleteSpace(int customerId, int spaceId) : IRequest<ResultMessage<DeleteResult>>
 {
-    public DeleteSpace(int customerId, int spaceId)
-    {
-        CustomerId = customerId;
-        SpaceId = spaceId;
-    }
-    
-    public int CustomerId { get; }
-    public int SpaceId { get; }
+    public int CustomerId { get; } = customerId;
+    public int SpaceId { get; } = spaceId;
 }
 
-public class DeleteSpaceHandler : IRequestHandler<DeleteSpace, ResultMessage<DeleteResult>>
+public class DeleteSpaceHandler(
+    ISpaceRepository spaceRepository,
+    ILogger<DeleteSpaceHandler> logger)
+    : IRequestHandler<DeleteSpace, ResultMessage<DeleteResult>>
 {
-    private readonly ISpaceRepository spaceRepository;
-    private readonly ILogger<DeleteSpaceHandler> logger;
-
-    public DeleteSpaceHandler(
-        ISpaceRepository spaceRepository,
-        ILogger<DeleteSpaceHandler> logger)
-    {
-        this.spaceRepository = spaceRepository;
-        this.logger = logger;
-    }
-    
     public async Task<ResultMessage<DeleteResult>> Handle(DeleteSpace request, CancellationToken cancellationToken)
     {
+        if (request.SpaceId == 0)
+        {
+            return new ResultMessage<DeleteResult>("Space 0 cannot be deleted", DeleteResult.Conflict);
+        }
+
         logger.LogDebug("Deleting Space {SpaceId}", request.SpaceId);
         var deleteResult = await spaceRepository.DeleteSpace(request.CustomerId, request.SpaceId, cancellationToken);
